% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_trend.R
\name{ts_trend}
\alias{ts_trend}
\title{Trend Extraction by Moving Averages}
\usage{
ts_trend(
  x,
  order = 5,
  plot = T,
  detrend = T,
  detrendPlot = F,
  type = "additive"
)
}
\arguments{
\item{x}{A time series}

\item{order}{Order of moving average smoother. Defaults to 5.}

\item{plot}{Whether to plot the trend or not. Defaults to TRUE.}

\item{detrend}{Whether to remove the calculated trend from the time series or not. TRUE needed to use with ts_season(). Defaults to TRUE.}

\item{detrendPlot}{Whether to plot the detrended series. Defaults to FALSE.}
}
\value{
If either plot or detrendPlot a list containing the series and the plots, otherwise a list containing the trend and the order of the moving average.
}
\description{
Extracts and plots the trend for a time series.
}
\details{
The moving average smoother averages the nearest order periods of each observation.
}
\examples{
 x <- AirPassengers
 ts_trend(x)
}
\seealso{
ma()
}
\author{
David Alberto Mateos Montes de Oca
}
\keyword{timeseries,}
\keyword{trend}
\keyword{ts,}
